Refactored restaurant gallery logic into separate component
